(tests): migrate all tests to api_test.go files
