--- fastly/airuntimecontrol/v1/provider/api_validation_test.go
@@ -1,17 +1,0 @@
-package provider
-
-import (
-	"context"
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
-	"github.com/fastly/go-fastly/v17/fastly"
-)
-
-func TestClient_ListModels_validation(t *testing.T) {
-	ctx := context.TODO()
-
-	_, err := ListModels(ctx, fastly.TestClient, &ListModelsInput{ProviderID: nil})
-	require.ErrorIs(t, err, fastly.ErrMissingProviderID)
-}
